// src/app/shared/model/pages.model.ts
export interface formFields {
  name: string;
  inputType: string;
  label: string;
  minLength: number;
  maxLength: number;
  placeholder: string;
  required: boolean;
  status: string;
  errorMessages: any;
  uuid: string;
}

export interface userLevels {
  level: number;
  title: string;
  uuid: string;
  shortName: string;
  canUserBeCreated : boolean
}

export interface modules {
  name: string;
  shortName: string;
  status: number;
  addedBy: string;
  parentId: string | null;
  uuid: string;
  parentName: string | null;
  categoryNames: any[];
}